refactor: remove useCallback in main page
