+++ src/components/main/index.jsx
@@ -1,91 +1,79 @@
-import React, { useCallback, useEffect, useRef } from 'react';
+import React, { useEffect, useRef } from 'react';
 import styled from 'styled-components';
 import { toast, ToastContainer } from 'react-toastify';
 import { injectStyle } from 'react-toastify/dist/inject-style';
 import { useDispatch, useSelector } from 'react-redux';
 import { actions, selector } from '../../store/modules';
-import UserRegister from './userRegister';
 import { reducerState } from '../../utils/reducer';
 import { PageLayout, Button } from '../common';
-import Loading from '../common/loading';
+import Loading from '../common/Loading';
+import UserName from './UserName';
+import UserGender from './UserGender';
+import { missingItems } from '../../constants/user';
 
 const MainPage = () => {
   const dispatch = useDispatch();
   const nameRef = useRef(null);
   const genderRef = useRef(null);
-  const userName = useSelector(selector.getUserName);
-  const userGender = useSelector(selector.getUserGender);
   const isQuestionLoading = useSelector(selector.isQuestionLoading);
+  const missingItem = useSelector(selector.getMissingItem);
 
   useEffect(() => injectStyle(), []);
 
-  const nameHandler = useCallback(
-    (name) => dispatch(actions.saveName(name)),
-    [dispatch],
-  );
-  const genderHandler = useCallback(
-    (gender) => dispatch(actions.saveGender(gender)),
-    [dispatch],
-  );
+  const handleSubmit = (e) => {
+    e.preventDefault();
 
-  const handleSubmit = useCallback(
-    (e) => {
-      e.preventDefault();
-      if (!userName) {
-        nameRef.current.focus();
-        toast.error('한글 2자 이상 입력해주세요.', {
-          toastId: 'name-error-id',
-        });
-        return;
-      }
+    if (missingItem === missingItems.name) {
+      nameRef.current.focus();
+      toast.error('한글 2자 이상 입력해주세요.', {
+        toastId: 'user-name-error',
+      });
+      return;
+    }
 
-      if (!userGender) {
-        genderRef.current.focus();
-        toast.error('성별은 필수 항목입니다.', {
-          toastId: 'gender-error-id',
-        });
-        return;
-      }
+    if (missingItem === missingItems.gender) {
+      genderRef.current.focus();
+      toast.error('성별을 선택해주세요.', {
+        toastId: 'user-gender-error',
+      });
+      return;
+    }
 
-      dispatch(actions.reqQuestions(reducerState.loading()));
-    },
-    [dispatch, userGender, userName],
-  );
+    dispatch(actions.reqQuestions(reducerState.loading()));
+  };
+
+  if (isQuestionLoading) {
+    return <Loading />;
+  }
 
   return (
     <PageLayout
       main={
-        <>
-          {isQuestionLoading && <Loading />}
-          <section>
-            <StyledMainTitle>직업가치관검사</StyledMainTitle>
-            <StyledUserContainer onSubmit={handleSubmit}>
-              <UserRegister
-                nameRef={nameRef}
-                userName={userName}
-                nameHandler={nameHandler}
-                genderRef={genderRef}
-                userGender={userGender}
-                genderHandler={genderHandler}
-              />
-              <Button type="submit">검사 시작</Button>
-              <ToastContainer
-                position="top-center"
-                autoClose={1000}
-                hideProgressBar={false}
-                pauseOnHover={false}
-                draggable={false}
-                theme="dark"
-              />
-            </StyledUserContainer>
-          </section>
-        </>
+        <section>
+          <StyledTitle>직업가치관검사</StyledTitle>
+          <StyledForm onSubmit={handleSubmit}>
+            <StyledFieldset>
+              <legend>사용자 등록</legend>
+              <UserName ref={nameRef} />
+              <UserGender ref={genderRef} />
+            </StyledFieldset>
+            <Button type="submit">검사 시작</Button>
+            <ToastContainer
+              position="top-center"
+              autoClose={1000}
+              hideProgressBar={false}
+              pauseOnHover={false}
+              draggable={false}
+              theme="dark"
+            />
+          </StyledForm>
+        </section>
       }
     />
   );
 };
 
-const StyledMainTitle = styled.h1`
+const StyledTitle = styled.h1`
   color: white;
   font-size: 2.2rem;
   margin-bottom: 2vh;
@@ -95,11 +83,20 @@ const StyledMainTitle = styled.h1`
   }
 `;
 
-const StyledUserContainer = styled.form`
+const StyledForm = styled.form`
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+`;
+
+const StyledFieldset = styled.fieldset`
+  border: none;
+  margin-bottom: 2vh;
+
+  > legend {
+    font-size: 0;
+  }
 `;
 
 export default MainPage;
