+++ src/constants/user/index.js
@@ -1,0 +1,1 @@
+export const missingItems = Object.freeze({ name: 0, gender: 1, not: 2 });
